extract file upload status handler
